:sparkles: Add a sort list of candidates route
This route takes in a set of tags with a weight percentage on each tag. These tags are then compared to each candidate in the list of the category to see how many matches there are.

A candidate that has more matched tags is ranked higher than a candidate that has the least matched tags.